pyproject.toml
[project]
name = "text-to-cad"
version = "0.1.0"
description = "Convert natural language queries to CAD design. The CAD design is just few prompts away!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.0.4",
    "cadquery>=2.5.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "langchain-openai>=1.1.3",
    "langchain>=1.2.0",
    "langchain-chroma>=1.1.0",
    "pypdf>=6.5.0",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.0",
    "tqdm>=4.67.1",
    "pyvista>=0.46.4",
]

[dependency-groups]
cdk = [
    "aws-cdk-lib>=2.230.0",
    "constructs>=10.4.3",
    "pyyaml>=6.0.3",
]
dev = [
    "rich>=14.2.0",
]
